PLAT-305 -pr suggestions - moved Column class to column.py
